:truck: 경사면 로직들을 SlopeComponent 로 분리
* 필요한 곳에서 Public Slope State 로 가져와서 쓰는 구조
