+++ Runelight-Smash-2021/Assets/Scripts/Component.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e5711423e3e40304a9f982b89ac44da2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
