+++ Runelight-Smash-2021/Assets/Scripts/Component/SlopeComponent.cs.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 7f45e4acb7a8cee4eb91fe97d8476438
+guid: c3c6a524bab19044c91693cb39279ea4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
-  executionOrder: 200
+  executionOrder: 100
   icon: {instanceID: 0}
   userData: 
   assetBundleName:
